% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{general_sqrt}
\alias{general_sqrt}
\title{Generalized Square Roots}
\usage{
general_sqrt(x)
}
\arguments{
\item{x}{A real number - integers or doubles are both acceptable.
Vectors must be of length 1.}
}
\value{
A complex value. If x is positive, the imaginary component is equal to 0;
if x is negative, the real component is equal to 0.
}
\description{
\code{general_sqrt} returns the square root of any real number
}
\examples{
general_sqrt(10)
general_sqrt(-10)
general_sqrt(-1)

}
