% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loss}
\alias{loss}
\title{Squared Difference Function}
\usage{
loss(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{Another numeric vector, of the same length as x}
}
\value{
A numeric vector of the same length as x and y
}
\description{
\code{loss} returns the squard difference element-wise between two vectors
}
\examples{
loss(c(1,2,3), c(2,3,4))

}
