% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{luxspending}
\alias{luxspending}
\title{Luxury spending of 100 individuals}
\format{A data frame with 100 rows and 2 columns.
\describe{
  \item{Income}{Total Income, in USD}
  \item{Luxury_goods}{Spending on luxury goods, in USD}
}}
\source{
This dataset was randomly generated
and is intended only to be used as an example for package
development and the scalep function.
}
\usage{
luxspending
}
\description{
This dataset lists the net incomes and the luxury spending
of 100 individuals.
}
\keyword{datasets}
