% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaledproportion.R
\name{scalep}
\alias{scalep}
\title{Scaled Proportion of Two Column Dataframe}
\usage{
scalep(d, x = 1)
}
\arguments{
\item{d}{Two column dataframe, filled with numeric values}

\item{x}{Optional scaling constant, by default set to 1}
}
\value{
A numeric vector of the same length as the columns of the Dataframe.
}
\description{
\code{scalep} returns the second column divided element-wise by the first, scaled (multiplied) by an optional scalar.
It will also graph a histogram of the result using ggplot2's qplot function,
or the default histogram function if ggplot2 is unavailable.
}
\examples{
income <- c(30000, 40000, 20000, 50000)
spending <- c(3000, 4000, 500, 100)
d <- data.frame(income, spending)
scalep(d, x=3)

}
